Handle trailing comma in render calls
There is a final false in argument_nodes which is typicaly skiped,
but when there is a trailing comma and you are parsing with ripper, instead of a trailing comma, ripper returns the arguement node we want in the last argument.

The same structure appears when the last argument is a proc like `&proc`.
This keeps that last argument in cases where there is a trailing comma or a last argument proc.